pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["Reza"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.3"
model-bakery = "^1.17.0"


[tool.poetry.group.dev.dependencies]
model-bakery = "^1.17.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
